+++ lib/connection.dart
@@ -1,0 +1,5 @@
+export 'package:flutter_bloc_patterns/src/tools/connection/connection.dart';
+export 'package:flutter_bloc_patterns/src/tools/connection/connection_bloc.dart';
+export 'package:flutter_bloc_patterns/src/tools/connection/connection_builder.dart';
+export 'package:flutter_bloc_patterns/src/tools/connection/connection_listener.dart';
+export 'package:flutter_bloc_patterns/src/tools/connection/connection_repository.dart';