+++ lib/src/tools/connection/connection.dart
@@ -1,0 +1,2 @@
+/// The Internet connection state. The device can be either [online] or [offline].
+enum Connection { online, offline }