+++ test/src/tools/connection/connection_test_data.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_bloc_patterns/src/tools/connection/connection.dart';
+
+const online = Connection.online;
+const offline = Connection.offline;
+const noChanges = Stream<Connection>.empty();