+++ test/src/util/bloc_state_assertion.dart
@@ -1,16 +1,15 @@
 import 'package:bloc/bloc.dart';
-import 'package:flutter_bloc_patterns/src/view/view_state.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-BlocStateAssertion withBloc(Bloc<dynamic, ViewState> bloc) =>
+BlocStateAssertion<S> withBloc<S>(BlocBase<S> bloc) =>
     BlocStateAssertion._internal(bloc);
 
-class BlocStateAssertion {
-  final Bloc<dynamic, ViewState> _bloc;
+class BlocStateAssertion<S> {
+  final BlocBase<S> _bloc;
 
   BlocStateAssertion._internal(this._bloc);
 
-  Future<void> expectStates(Iterable<ViewState> states) {
+  Future<void> expectStates(Iterable<S> states) {
     return expectLater(
       _bloc.stream,
       emitsInOrder(states),