+++ test/src/view/view_state_listener_test.dart
@@ -153,4 +153,8 @@ void main() {
     verifyNever(() => emptyCallback.call(any()));
     verify(() => errorCallback.call(any(), _someException));
   });
+
+  tearDown(() {
+    bloc.close();
+  });
 }