+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.11.0]
+
+* Migrating to `bloc` 8.1.x and `flutter_bloc` 8.1.x,
+* `ConnectionBloc` - a BLoC that exposes the Internet connection state to the UI.
+
 ## [0.10.0]
 
 * Migrating to `bloc` 8.0.x and `flutter_bloc` 8.0.x,